+++ apps/imapi/src/websocket_handler.erl
@@ -28,8 +28,8 @@ init(Req0, State0) ->
              num_acceptors => infinity,
              max_connections => infinity,
              max_frame_size => 1048576,  % 1MB
-             % Cowboy关闭连接空闲120秒 默认值为 60000
-             idle_timeout => 120000
+             % Cowboy关闭连接空闲128秒 默认值为 60000
+             idle_timeout => 128000
             },
     State1 = State0#{dtype => DType, did => DID},
     case throttle:check(throttle_ws, DID) of